chore(test/browser): extract runVitest helper
This is in preparation for adding a new test file which will run the
tests with config.base set to confirm the fix for #4686.
