+++ test/browser/specs/runner.test.mjs
@@ -1,31 +1,14 @@
 import assert from 'node:assert'
-import { readFile } from 'node:fs/promises'
 import test from 'node:test'
-import { execa } from 'execa'
+import runVitest from './run-vitest.mjs'
 
-const browser = process.env.BROWSER || (process.env.PROVIDER === 'playwright' ? 'chromium' : 'chrome')
-const argv = ['vitest', '--run', `--browser.name=${browser}`]
-
-if (browser !== 'safari')
-  argv.push('--browser.headless')
-
-const { stderr, stdout } = await execa('npx', argv, {
-  env: {
-    ...process.env,
-    CI: 'true',
-    NO_COLOR: 'true',
-  },
-  reject: false,
-})
-
-const browserResult = await readFile('./browser.json', 'utf-8')
-const browserResultJson = JSON.parse(browserResult)
-
-const getPassed = results => results.filter(result => result.status === 'passed')
-const getFailed = results => results.filter(result => result.status === 'failed')
-
-const passedTests = getPassed(browserResultJson.testResults)
-const failedTests = getFailed(browserResultJson.testResults)
+const {
+  stderr,
+  stdout,
+  browserResultJson,
+  passedTests,
+  failedTests,
+} = await runVitest()
 
 await test('tests are actually running', async () => {
   assert.ok(browserResultJson.testResults.length === 10, 'Not all the tests have been run')
